+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Email templates for Likes and Reposts
 * Improve Interactions moderation
 * Compatibility with Akismet
+* Comment type mapping for `Like` and `Announce`
 
 ## [4.4.0] - 2024-12-09